pyproject.toml
[project]
name = "xiaomi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "cachetools>=6.2.5",
    "charset-normalizer",
    "colorama",
    "fastapi>=0.128.0",
    "pillow",
    "pycryptodome",
    "pydantic>=2.12.5",
    "python-miio",
    "pyyaml>=6.0.3",
    "requests",
    "tabulate>=0.9.0",
    "tenacity>=9.1.2",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
]
